Move runGit into panes.ts, and out of palui's host
A one-function follow-on to the pane layer that pays for itself twice.

runGit belongs with a session's lifecycle: it acts on that session's workdir
and, when the action finishes, re-reads exactly the two things panes.ts
already owns (refreshSessionStats for the new ahead/behind, refreshBranch for
the label). Every callee it has was already in that module.

And it lets palui.ts import it instead of taking it as a host member, so the
palette's host drops from eleven callees to ten — the same seam-rule-1 saving
the ${input:…} move got. Cycle check re-run after the new import: still no
cycles across 33 modules.

Verified mechanical: the relocated block diffed against
`git show HEAD:src/main.ts` differs by one `export` keyword.

main.ts 730 -> 693. Gates: tsc --noEmit clean (exit 0), 368 vitest pass. No
Rust touched.
